Moved models back to serialization project
As fileview is "internal" accepts irsdkSharp.Serialization as a "friend"

The only difference is based on accembly and removing serialization from base project. this should have no effect on serialisation speed.

will need to look at what happens on merge as some items have been changed in irsdk for some of the work I did and Chris did so merge is going to be fun!
