+++ src/irsdkSharp.Serialization/Enums/Fastest/PaceMode.cs
@@ -1,0 +1,13 @@
+﻿using System;
+
+namespace irsdkSharp.Serialization.Enums.Fastest
+{
+    public enum PaceMode
+    {
+        SingleFileStart = 0,
+        DoubleFileStart,
+        SingleFileRestart,
+        DoubleFileRestart,
+        NotPacing,
+    }
+}
