+++ src/irsdkSharp.Serialization/Enums/Fastest/TrackSurface.cs
@@ -1,0 +1,11 @@
+﻿namespace irsdkSharp.Serialization.Enums.Fastest
+{
+    public enum TrackSurface
+    {
+        NotInWorld = -1,
+        OffTrack,
+        InPitStall,
+        AproachingPits,
+        OnTrack
+    }
+}
