--- src/irsdkSharp/AccessorExtensions.cs
@@ -1,32 +1,0 @@
-﻿using System.IO.MemoryMappedFiles;
-using System.Text;
-
-namespace irsdkSharp
-{
-    public static class AccessorExtensions
-    {
-        public static string ReadString(this MemoryMappedViewAccessor accessor, int offset, int maxLength, int minLength = 0)
-        {
-            StringBuilder sb;
-            if (minLength > 0)
-            {
-                sb = new(minLength);
-            }
-            else
-            {
-                sb = new();
-            }
-            char c;
-            for (int i = 0; i < maxLength; i++)
-            {
-                c = (char)accessor.ReadByte(offset + i);
-                if(c == '\0')
-                {
-                    break;
-                }
-                sb.Append(c);
-            }
-            return sb.ToString();    
-        }
-    }
-}
