--- src/irsdkSharp/Enums/PaceMode.cs
@@ -1,13 +1,0 @@
-﻿using System;
-
-namespace irsdkSharp.Enums
-{
-    public enum PaceMode
-    {
-        SingleFileStart = 0,
-        DoubleFileStart,
-        SingleFileRestart,
-        DoubleFileRestart,
-        NotPacing,
-    }
-}
