--- src/irsdkSharp/Enums/TrackSurface.cs
@@ -1,11 +1,0 @@
-﻿namespace irsdkSharp.Enums
-{
-    public enum TrackSurface
-    {
-        NotInWorld = -1,
-        OffTrack,
-        InPitStall,
-        AproachingPits,
-        OnTrack
-    }
-}
