--- src/irsdkSharp/Enums/TrackSurfaceMaterial.cs
@@ -1,37 +1,0 @@
-﻿namespace irsdkSharp.Enums
-{
-    public enum TrackSurfaceMaterial
-    {
-        SurfaceNotInWorld = -1,
-        UndefinedMaterial = 0,
-
-        Asphalt1Material,
-        Asphalt2Material,
-        Asphalt3Material,
-        Asphalt4Material,
-        Concrete1Material,
-        Concrete2Material,
-        RacingDirt1Material,
-        RacingDirt2Material,
-        Paint1Material,
-        Paint2Material,
-        Rumble1Material,
-        Rumble2Material,
-        Rumble3Material,
-        Rumble4Material,
-
-        Grass1Material,
-        Grass2Material,
-        Grass3Material,
-        Grass4Material,
-        Dirt1Material,
-        Dirt2Material,
-        Dirt3Material,
-        Dirt4Material,
-        SandMaterial,
-        Gravel1Material,
-        Gravel2Material,
-        GrasscreteMaterial,
-        AstroturfMaterial,
-    }
-}
