--- src/irsdkSharp/Models/CameraInfo/CameraInfoModel.cs
@@ -1,9 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace irsdkSharp.Models
-{
-    public class CameraInfoModel
-    {
-        public List<GroupModel> Groups { get; set; }
-    }
-}
