--- src/irsdkSharp/Models/CameraInfo/CameraModel.cs
@@ -1,8 +1,0 @@
-﻿namespace irsdkSharp.Models
-{
-    public class CameraModel
-    {
-        public int CameraNum { get; set; }// %d
-        public string CameraName { get; set; }// %s
-    }
-}
