--- src/irsdkSharp/Models/QualifyResultsInfo/ResultModel.cs
@@ -1,11 +1,0 @@
-﻿namespace irsdkSharp.Models
-{
-    public class ResultModel
-    {
-        public int Position { get; set; }// %d
-        public int ClassPosition { get; set; }// %d
-        public int CarIdx { get; set; }// %d
-        public int FastestLap { get; set; }// %d
-        public float FastestTime { get; set; }// %.3f
-    }
-}
