--- src/irsdkSharp/Models/RadioInfo/RadioInfoModel.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace irsdkSharp.Models
-{
-    public class RadioInfoModel
-    {
-        public int SelectedRadioNum { get; set; }
-        public List<RadioModel> Radios { get; set; }
-    }
-}
