--- src/irsdkSharp/Models/SessionInfo/FastestLapModel.cs
@@ -1,9 +1,0 @@
-﻿namespace irsdkSharp.Models
-{
-    public class FastestLapModel
-    {
-        public int CarIdx { get; set; }// %d
-        public int FastestLap { get; set; }// %d
-        public float FastestTime { get; set; }// %.3f
-    }
-}
