--- src/irsdkSharp/Models/SplitTimeInfo/SectorModel.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace irsdkSharp.Models
-{
-    public class SectorModel
-    {
-        public int SectorNum { get; set; }// %d
-        public float SectorStartPct { get; set; }// %.3f
-    }
-}
