--- src/irsdkSharp/Models/SplitTimeInfo/SplitTimeInfoModel.cs
@@ -1,9 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace irsdkSharp.Models
-{
-    public class SplitTimeInfoModel
-    {
-        public List<SectorModel> Sectors { get; set; }
-    }
-}
