--- src/irsdkSharp/Models/WeekendInfo/TelemetryOptionsModel.cs
@@ -1,7 +1,0 @@
-﻿namespace irsdkSharp.Models
-{
-    public class TelemetryOptionsModel
-    {
-        public string TelemetryDiskFile { get; set; }
-    }
-}
