--- src/irsdkSharp/Models/WeekendInfo/WeekendOptionsModel.cs
@@ -1,35 +1,0 @@
-﻿namespace irsdkSharp.Models
-{
-    public class WeekendOptionsModel
-    {
-        public int NumStarters { get; set; }
-        public string StartingGrid { get; set; }
-        public string QualifyScoring { get; set; }
-        public string CourseCautions { get; set; }
-        public int StandingStart { get; set; }
-        public string Restarts { get; set; }
-        public string WeatherType { get; set; }
-        public string Skies { get; set; }
-        public string WindDirection { get; set; }
-        public string WindSpeed { get; set; }
-        public string WeatherTemp { get; set; }
-        public string RelativeHumidity { get; set; }
-        public string FogLevel { get; set; }
-        public int Unofficial { get; set; }
-        public string CommercialMode { get; set; }
-        public string NightMode { get; set; }
-        public int IsFixedSetup { get; set; }
-        public string StrictLapsChecking { get; set; }
-        public int HapSOpenRegistration { get; set; }
-        public int HardcoreLevel { get; set; }
-        public string ShortParadeLap { get; set; }
-        public string TimeOfDay { get; set; }
-        public string Date { get; set; }
-        public string EarthRotationSpeedupFactor { get; set; }
-        public string HasOpenRegistration { get; set; }
-        public int NumJokerLaps { get; set; }
-        public string IncidentLimit { get; set; }
-        public string FastRepairsLimit { get; set; }
-        public int GreenWhiteCheckeredLimit { get; set; }
-    }
-}
